refactor: file four types where they belong
Each of these was imported from a module that had nothing to do with why
it was wanted, and three of them created a dependency edge that reflected
no real dependency.

`ancestors` was in `pick`, though it is a plain `ChildOf` walk. It is the
upward counterpart to `traversal::walk` — what a pick resolves *to* (click
target, focus path, hover chain) is decided by going up from the one
entity picking returned — so it belongs beside it. `focus`'s only use of
`pick` was this function, so that edge is now gone entirely.

`RectGeometry` was in `color_rect`, though its `Layout` impl is exactly
"resolve `Sizing` around a declared content size" and six widgets use it.
It made `color_rect` the most depended-on module in the crate for a reason
unrelated to coloured rectangles.

`AlignItems` was in `layout`, which made `layout` and `sizing` depend on
each other in both directions: containers read it, and `align_self` — a
per-item override, and so a `sizing` concern — names the same type. Filing
the shared vocabulary with the item side breaks the cycle. `layout`
re-exports it, so callers are unaffected.

`Easing` was in `animation`, so `interaction` depended on the fade plugin
for a curve. Both are independently registrable and neither owns easing.
